requirements.txt
streamlit
firebase_admin
openai
re
json
google.oauth2
